fix(actions): bound array index in float space before int narrowing (CodeQL)
exprIndex compared int(n) against len(arr) after the float-to-int
conversion, so a float beyond the int range would overflow the cast and
could pass the bound check. Guard 0 <= n < len(arr) in float space first;
the int conversion is then in range by construction. This is the real
form of the deviation the SARIF allowlist had adjudicated, so its entry
(stale-pathed since expressions.go moved to internal/actions in #259) is
removed.

+++ internal/actions/expressions.go
@@ -751,7 +751,11 @@ func exprFilterValues(v interface{}) interface{} {
 func exprIndex(v, idx interface{}) interface{} {
 	if arr, ok := v.([]interface{}); ok {
 		n := exprToNumber(idx)
-		if math.IsNaN(n) || n < 0 || int(n) >= len(arr) {
+		// Bound n as a float before narrowing to int: a float larger than the
+		// int range would overflow the conversion, so comparing int(n) after
+		// the cast is unsafe. Once 0 <= n < len(arr) holds in float space, the
+		// int conversion is in range by construction.
+		if math.IsNaN(n) || n < 0 || n >= float64(len(arr)) {
 			return nil
 		}
 		return arr[int(n)]
